LPS-44962 move
