+++ modules/portal/portal-sso-token/src/com/liferay/portal/sso/token/auth/CookieTokenRetriever.java
@@ -15,8 +15,6 @@
 package com.liferay.portal.sso.token.auth;
 
 import com.liferay.portal.kernel.util.CookieKeys;
-import com.liferay.portal.sso.token.spi.TokenLocation;
-import com.liferay.portal.sso.token.spi.TokenRetriever;
 
 import javax.servlet.http.HttpServletRequest;
