+++ modules/portal/portal-sso-token/src/com/liferay/portal/sso/token/auth/TokenRetriever.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.sso.token.spi;
+package com.liferay.portal.sso.token.auth;
 
 import javax.servlet.http.HttpServletRequest;
